+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Added new `Type.tuple(...)` type.
 * Rewrite of `either` to support any amount of specs, instead of the maximum of 9 from before.
 * Added schema attribute descriptions to `Type.object` and `Type.interface` definitions.
+* Added `Type.literalValue` for literal constant values.
 * TODO!!!
 
 ## v0.4.1